change data from int to double. + fix radix which magically worked when it shouldn't have.

+++ Headers/bogoSort.h
@@ -1,6 +1,6 @@
 #include<iostream>
 #include <vector>
 
-void bogoSort(std::vector<int>& v){
+void bogoSort(std::vector<double>& v){
 
 }
