+++ Headers/heapSort.h
@@ -1,6 +1,8 @@
 #include<iostream>
 #include <vector>
 
-void heapSort(std::vector<int>& v){
+void heapSort(std::vector<double>& v){
+    ///priority_queue<double> pq;
+
 
 }
